Move LocalReplicaChanges type
